[Fix - GlobalCollect TLS support]
This fixes  a bug in GlobalCollect SDK: SSL Connection to  https://api-preprod.globalcollect.com/v1 was broken. ( during handshake phase)
GC uses the wrong implementation of HttpClient that ignores the https.protocols which allow you to overide the supported protocols it in JDK7.
By default, the default HTTPClient will use TLSV1 which the default in JDK7

This implementation:

- Allows connection to GC API with a supported TLS version. TLSV1.1 will be used by default instead of TLSV1
- TLSV1 support is being removed by GC API and it causes troubles with jre6/7. (TLSV1 = default in jdk7 jdk6. TLSV1.2 is not supported in jdk6)
- Support of TLSV1.1, TLSV1.2, there are still supported today. ( May 2018 )
- You can override this by setting the standard "https.protocols" parameter at startup to load supported your own supported https protocols